LayoutTests/fast/events/scale-and-scroll-body-expected.txt

PASS window.document.body.scrollTop is 30
PASS window.document.body.scrollLeft is 30
PASS window.document.body.scrollTop is 100
PASS window.document.body.scrollLeft is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
